refactor(node/url): Move toPathIfFileURL to internal
